save file

+++ 18_Solution.cpp
@@ -14,13 +14,13 @@ public:
         int length = nums.size();
         for(int i = 0; i < length - 3; ++i){
             if(i > 0 && nums[i] == nums[i - 1]) continue;
-            if(nums[i] + nums[i+1] + nums[i+2] + nums[i+3] > target) break;
-            if(nums[i] + nums[length - 3] + nums[length - 2] + nums[length - 1] < target) continue;
+            if((long)nums[i] + nums[i+1] + nums[i+2] + nums[i+3] > target) break;
+            if((long)nums[i] + nums[length - 3] + nums[length - 2] + nums[length - 1] < target) continue;
 
             for(int j = i + 1; j < length -2; ++j){
                 if(j > i + 1 && nums[j] == nums[j-1]) continue;
-                if(nums[i] + nums[j] + nums[j + 1] + nums[j + 2] > target) break;
-                if(nums[i] + nums[j] + nums[length - 2] + nums[length - 1] < target) continue;
+                if((long)nums[i] + nums[j] + nums[j + 1] + nums[j + 2] > target) break;
+                if((long)nums[i] + nums[j] + nums[length - 2] + nums[length - 1] < target) continue;
                 
                 int left = j + 1, right = length - 1;
                 while(left < right){
